[#40] feat: apply MediaQueryWrapper
